requirements.txt
asgiref==3.7.2
Django==4.2.2
django-multiselectfield==0.1.12
djangorestframework==3.14.0
mysqlclient==2.2.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
